requirements.txt
Flask==3.0.3
requests==2.32.3
certifi==2024.8.30
qdrant-client==1.11.3
sentence-transformers==3.1.1
langchain-community==0.3.2
python-dotenv==1.0.1
PyPDF2==3.0.1
regex==2024.9.11
mysql-connector-python==9.0.0